fix undefined index status

+++ app/Models/Pharmacy/PrescriptionsDispensings.php
@@ -22,10 +22,10 @@ class PrescriptionsDispensings extends Model
     public static function boot() {
         static::saved(function($model) {
             $original = $model->getOriginal();
-            if ($model->status == 'dispensed' && $original["status"] != 'dispensed') {
+            if ($model->status == 'dispensed' && (!$original["status"] || $original["status"] != 'dispensed')) {
                 $model->createStockCard();
             }
-            if ($original['status']=='dispensed' && $model->status!='dispensed') {
+            if ($original["status"] && $original['status']=='dispensed' && $model->status!='dispensed') {
                 $model->StocksCards()->delete();
             }
         });
